# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/colors": {
      "get": {
        "operationId": "ColorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ფერების სია"
          }
        },
        "summary": "ფერების სია",
        "tags": [
          "colors"
        ]
      },
      "post": {
        "operationId": "ColorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateColorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ფერი შეიქმნა"
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ახალი ფერის შექმნა (ADMIN)",
        "tags": [
          "colors"
        ]
      }
    },
    "/colors/{id}": {
      "get": {
        "operationId": "ColorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ფერი"
          },
          "404": {
            "description": "ფერი ვერ მოიძებნა"
          }
        },
        "summary": "კონკრეტული ფერის მიღება",
        "tags": [
          "colors"
        ]
      },
      "put": {
        "operationId": "ColorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateColorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ფერი განახლდა"
          },
          "404": {
            "description": "ფერი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ფერის განახლება (ADMIN)",
        "tags": [
          "colors"
        ]
      },
      "delete": {
        "operationId": "ColorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ფერი წაიშალა"
          },
          "404": {
            "description": "ფერი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ფერის წაშლა (ADMIN)",
        "tags": [
          "colors"
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "გვერდის ნომერი (იწყება 1-დან)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "ჩანაწერების რაოდენობა თითო გვერდზე",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "დალაგების ველი",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "დალაგების მიმართულება",
            "schema": {
              "default": "DESC",
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "საძიებო ტექსტი — ეძებს name და description ველებში",
            "schema": {
              "example": "ყურსასმენი",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "გაფილტვრა კატეგორიის მიხედვით",
            "schema": {
              "format": "uuid",
              "example": "e3b0c442-98fc-1c14-9afc-2c963f66afa6",
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "მინიმალური ფასი",
            "schema": {
              "minimum": 0,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "მაქსიმალური ფასი",
            "schema": {
              "minimum": 0,
              "example": 500,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "გაფილტვრა აქტიურობის მიხედვით",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasDiscount",
            "required": false,
            "in": "query",
            "description": "გაფილტვრა ფასდაკლების მიხედვით — true: მხოლოდ ფასდაკლებიანი, false: მხოლოდ ფასდაკლების გარეშე",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "discountPercent",
            "required": false,
            "in": "query",
            "description": "ზუსტი ფასდაკლების პროცენტის მიხედვით გაფილტვრა",
            "schema": {
              "minimum": 0,
              "example": 15,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "პროდუქტების გვერდიანი სია"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტების სია — ძიება, ფილტრები, პაგინაცია. ტოკენის გარეშე/non-ADMIN მომხმარებელს ყოველთვის მხოლოდ isActive=true პროდუქტები უჩანს; ADMIN-ს (ვალიდური ბირერ ტოკენით) — ყველა, თუ isActive query param-ით სხვა არაა მოთხოვნილი",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "პროდუქტი შეიქმნა",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ახალი პროდუქტის შექმნა (ADMIN)",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "პროდუქტი"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "summary": "კონკრეტული პროდუქტის მიღება",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "პროდუქტი განახლდა",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის განახლება (ADMIN)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "პროდუქტი წაიშალა"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის წაშლა (ADMIN)",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/similar": {
      "get": {
        "operationId": "ProductsController_findSimilar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "მსგავსი პროდუქტები"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "summary": "მსგავსი პროდუქტების სია (პროდუქტის გვერდის სლაიდერისთვის) — იმავე კატეგორიის აქტიური პროდუქტები, საწყისის გამოკლებით",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/attribute-values": {
      "get": {
        "operationId": "ProductsController_getAttributeValues",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "attribute value-ები"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "summary": "პროდუქტის attribute value-ების სია",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_setAttributeValues",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProductAttributeValuesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "attribute value-ები განახლდა"
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის attribute value-ების bulk set (ADMIN) — მთლიანად ანაცვლებს არსებულს",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/additional-info": {
      "get": {
        "operationId": "ProductsController_getAdditionalInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "დამატებითი ინფორმაციის ბლოკები"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "summary": "პროდუქტის დამატებითი ინფორმაციის ბლოკების სია",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_addAdditionalInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductAdditionalInfoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ბლოკი შეიქმნა"
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტს დამატებითი ინფორმაციის ახალი ბლოკის დამატება (ADMIN)",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/additional-info/{infoId}": {
      "put": {
        "operationId": "ProductsController_updateAdditionalInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "infoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductAdditionalInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ბლოკი განახლდა"
          },
          "404": {
            "description": "ბლოკი ან პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის დამატებითი ინფორმაციის ბლოკის განახლება (ADMIN)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_removeAdditionalInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "infoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ბლოკი წაიშალა"
          },
          "404": {
            "description": "ბლოკი ან პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის დამატებითი ინფორმაციის ბლოკის წაშლა (ADMIN)",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/colors": {
      "get": {
        "operationId": "ProductsController_getColors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ფერები"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "summary": "პროდუქტზე მიბმული ფერების სია (stock-ითურთ)",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_setColors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProductColorsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ფერები განახლდა"
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "404": {
            "description": "პროდუქტი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პროდუქტის ფერების bulk set (ADMIN) — მთლიანად ანაცვლებს არსებულს",
        "tags": [
          "products"
        ]
      }
    },
...
    "/branches": {
      "get": {
        "operationId": "BranchesController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ფილიალების სია"
          }
        },
        "summary": "აქტიური ფილიალების სია (checkout-ისთვის)",
        "tags": [
          "branches"
        ]
      },
      "post": {
        "operationId": "BranchesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ფილიალი დაემატა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ახალი ფილიალის დამატება (ADMIN)",
        "tags": [
          "branches"
        ]
      }
    },
...
  },
  "info": {
    "title": "Online Shop API",
    "description": "REST API for the online shop backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateColorDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი სახელი — { ka: {name}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/NameTranslationsDto"
              }
            ]
          },
          "hexCode": {
            "type": "string",
            "pattern": "^#([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$",
            "description": "HEX კოდი frontend-ის სვოჩისთვის",
            "example": "#FF0000"
          }
        },
        "required": [
          "translations"
        ]
      },
      "UpdateColorDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი სახელი — { ka: {name}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/NameTranslationsDto"
              }
            ]
          },
          "hexCode": {
            "type": "string",
            "pattern": "^#([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$",
            "description": "HEX კოდი frontend-ის სვოჩისთვის",
            "example": "#FF0000"
          }
        }
      },
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "პროდუქტი წარმატებით შეიქმნა"
          },
          "data": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
    }
  }
}